migrate priority tagging methods
