+++ src/OrleansRuntime/Scheduler/PoliciedScheduler/SchedulingStrategies/SchedulerConstants.cs
@@ -1,0 +1,12 @@
+﻿
+
+namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
+{
+    internal static class SchedulerConstants
+    {
+        internal const long DEFAULT_PRIORITY = 0L;
+        internal const long DEFAULT_WIG_EXECUTION_COST = 0L;
+        internal const int DEFAULT_TASK_QUANTUM_MILLIS = 100;
+        internal const int DEFAULT_TASK_QUANTUM_NUM_TASKS = 0;
+    }
+}
